requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Bcrypt
python-dotenv
requests
Flask-WTF
email-validator
